pesticide_backend/requirements.txt
aioredis==1.3.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.4.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.1.0
bleach==3.1.4
Brlapi==0.6.6
certifi==2020.4.5.1
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
command-not-found==0.3
constantly==15.1.0
cryptography==3.0
cupshelpers==1.0
daphne==2.5.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filter==2.3.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-richtextfield==1.6
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
hiredis==1.1.0
httpie==2.1.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.6.0
importlib-resources==1.5.0
incremental==17.5.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.1
jupyterlab-server==1.1.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
notebook==6.0.3
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.0.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
pygobject==3.26.1
PyHamcrest==2.0.2
pylint==2.5.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.1.0
pyRFC3339==1.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2020.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.0
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.13.2
six==1.15.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
virtualenv==20.0.18
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
zope.interface==5.1.0
